requirements.txt
numpy
tqdm
torch==2.0.1+cu118
pandas
torchvision==0.15.2+cu118
Pillow==9
PyQt5
qt-material
matplotlib